Iterates scrollTop class
