+++ packages/ui/components/scrollTop/scripts.js
@@ -1,26 +1,27 @@
 import animateTo from '@runroom/purejs/lib/animateTo';
 import safeScrollTop from '@runroom/purejs/lib/safeScrollTop';
 
-const ELEMENT_ID = 'scroll-top';
-const scrollTopBtn = document.getElementById(ELEMENT_ID);
-let opts = {
-    showOffset: 0,
-    scrollOffset: 0,
-    speed: 300
+class ScrollTop {
+  constructor({ selector = '#scroll-top', showOffset = 0, scrollOffset = 0, speed = 300 }) {
+    this.scrollTopBtn = document.querySelector(selector);
+    this.showOffset = showOffset;
+    this.scrollOffset = scrollOffset;
+    this.speed = speed;
+
+    this._init();
+  }
+
+  _hideShowButton() {
+    const shouldBeVisible = safeScrollTop() > this.showOffset;
+    this.scrollTopBtn.classList[shouldBeVisible ? 'add' : 'remove']('is-active');
+  }
+
+  _init() {
+    this._hideShowButton();
+
+    this.scrollTopBtn.addEventListener('click', () => animateTo({ element: this.scrollOffset, speed: this.speed }));
+    window.addEventListener('scroll', () => this._hideShowButton());
+  }
 }
 
-const hideShowButton = () => {
-    const shouldBeVisible = safeScrollTop() > opts.showOffset;
-    scrollTopBtn.classList[shouldBeVisible ? 'add' : 'remove']('is-active');
-};
-
-const init = params => {
-    if (scrollTopBtn) {
-        opts = { ...opts, ...params };
-        hideShowButton();
-        scrollTopBtn.addEventListener('click', () => animateTo({ element: opts.scrollOffset, speed: opts.speed }));
-        window.addEventListener('scroll', () => hideShowButton());
-    }
-};
-
-init({ showOffset: 100 });
+const scrollTopButton = new ScrollTop({ showOffset: 100 });
